requirements.txt
requests==2.32.5
python-dotenv==1.2.1
PyJWT==2.10.1
streamlit==1.41.1

